move RotFlip to new file
